DAMMmodel

DAMMmodel.sDAMMmat

DAMMmodel.sDAMMmatq

DAMMmodel.DAMM

DAMMmodel.DAMMfit

DAMMmodel.DAMMmat

DAMMmodel.DAMMplot

DAMMmodel.DAMMviz

DAMMmodel.qbins

DAMMmodel.sDAMMmat — Type

struct sDAMMmat

Fields required to plot surface of DAMM

Fields

porosity::Float64

params::NTuple{7, Float64}

x::Vector{Float64}

y::Vector{Float64}

DAMM_Matrix::Matrix{Float64}

DAMMmodel.sDAMMmatq — Type

struct sDAMMmatq

Fields required to plot surface of DAMM

Fields

porosity::Float64

Tmed::Vector{Float64}

θmed::Vector{Float64}

Rmed::Vector{Float64}

params::NTuple{7, Float64}

x::Vector{Float64}

y::Vector{Float64}

DAMM_Matrix::Matrix{Float64}

DAMMmodel.DAMM — Method

DAMM(x::VecOrMat{<: Real}, p::NTuple{7, Float64})

Calculate respiration as a function of soil temperature and moisture.

Examples

julia> Tₛ = [18.0, 22.0] # soil temperature [°C]
julia> θ = [0.35, 0.22] # soil moisture [m³ m⁻³]
julia> x = hcat(Tₛ, θ)
julia> p = (1e9, 64.0, 3.46e-8, 2.0e-3, 0.4, 0.0125, 1.0) # αₛₓ, Eaₛₓ, kMₛₓ, kMₒ₂, Sxₜₒₜ, Q10kM
julia> DAMM(x, p)
  1.6 # μmolCO₂ m⁻² s⁻¹ 
  2.8 # μmolCO₂ m⁻² s⁻¹

DAMMmodel.DAMMfit — Method

DAMMfit(Ind_var, Resp, poro_val)

fit the DAMM model parameters to data.

Examples

julia> Tₛ = [19.0, 22.0] # soil temperature [°C]
julia> θ = [0.35, 0.22] # soil moisture [m³ m⁻³]
julia> Rₛ = [2.2, 2.8] # respiration observation
julia> Ind_var = hcat(Tₛ, θ)
julia> p = DAMMfit(Ind_var, Rₛ, 0.4) 
  10e8
  64
  9.6e-7
  1.3e-4
  0.4
  0.02
  1.0
julia> DAMM(Ind_var, p)
  2.2
  2.8

DAMMmodel.DAMMmat — Method

DAMMmat(Tₛ::Array{Float64, 1}, θ::Array{Float64, 1}, Rₛ::Array{Float64, 1}, r::Int64)

Generates a matrix of DAMM output for gridded inputs x and y Inputs: soil temperature (Tₛ), soil moisture (θ), respiration (Rₛ), resolution (r)

Examples:

julia> Tₛ = collect(15.0:2.5:40.0)
julia> θ = collect(0.2:0.05:0.7)
julia> Rₛ = [1.0, 1.2, 1.5, 2.0, 2.7, 3.8, 4.9, 6.7, 4.1, 2.0, 0.4]
julia> r = 10
julia> out = DAMMmat(Tₛ, θ, Rₛ, r)

DAMMmat(Tₛ::Array{Float64, 1}, θ::Array{Float64, 1}, Rₛ::Array{Float64, 1}, r::Int64, n::Int64)

Bin data by n quantiles

Examples:

julia> n = 4
julia> out = DAMMmat(Tₛ, θ, Rₛ, r, n)

DAMMmodel.DAMMplot — Method

DAMMplot(Tₛ::Array{Float64, 1}, θ::Array{Float64, 1}, Rₛ::Array{Float64, 1}, r::Int64)

Plot scatter of data and fitted DAMM surface

Example

julia> Tₛ = collect(15.0:2.5:40.0)
julia> θ = collect(0.2:0.05:0.7)
julia> Rₛ = [1.0, 1.2, 1.5, 2.0, 2.7, 3.8, 4.9, 6.7, 4.1, 2.0, 0.4]
julia> r = 10
julia> fig = DAMMplot(Tₛ, θ, Rₛ, r)

DAMMmodel.DAMMviz — Method

DAMMviz()

Interactive plot of the DAMM model

To open

julia> DAMMviz()

DAMMmodel.qbins — Method

qbins(x, y, z, n)

Bins x into n quantiles, each xbin into n quantiles of y, return z quantile

Examples

julia> df = DataFrame(x=1:20, y=6:25, z=11:30)
julia> xmed, ymed, zmed = qbins(df.T, df.M, df.R, 3)
  xmed = [9, 9, 9, 15, 15, 15, 21, 21, 21]
  ymed = [12, 14, 16, 19, 20.5, 22, 25, 27, 29]
  zmed = [2, 4, 6, 8.5, 10.5, 15, 17, 19]